bot/requirements.txt
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html
numpy==1.21.2
pillow==8.3.2
python-telegram-bot==13.10
torch==1.10.2+cpu
torchvision==0.11.3+cpu
scikit-image==0.19.1
gdown==4.2.1